links are generating and saving at redis
