+++ internal/model/model.go
@@ -1,41 +1,10 @@
 package model
 
-import (
-	"sync"
-)
 
 type Item struct {
-	Short string
-	Long  string
-	TTL   uint
-	Count uint
-}
-type Info struct {
-	mu   sync.RWMutex
-	link map[string]*Item
-}
-
-func (i *Info) Get(key string) *Item {
-	i.mu.RLock()
-	if val, ok := i.link[key]; ok {
-		return val
-	}
-	i.mu.RUnlock()
-	return nil
-}
-
-func (i *Info) Put(key string, item *Item) {
-	i.mu.Lock()
-	i.link[key] = item
-	i.mu.Unlock()
-}
-
-func (i *Info) Delete(key string) bool {
-	if _, ok := i.link[key]; ok {
-		i.mu.Lock()
-		delete(i.link, key)
-		i.mu.Unlock()
-		return true
-	}
-	return false
+	ShortLink  string
+	FullLink   string
+	UserID string
+	TTL    uint
+	Count  uint
 }
